NavBar
BookLetters is a tool that will create Book Newsletters. We tried to use
it to create static booklists, but needed a way for people to browse and
discover other booklists.

I created the following as a hack to the Book Letters Skin so that it
would add navigation.

First implemented 2012